+++ src/main/java/ru/practicum/shareit/item/repository/ItemRepository.java
@@ -9,9 +9,9 @@ import ru.practicum.shareit.item.model.Item;
 import java.util.List;
 
 @Repository
-public interface ItemRepository extends JpaRepository<Item, Integer> {
+public interface ItemRepository extends JpaRepository<Item, Long> {
 
-    List<Item> findByOwnerId(int ownerId, Pageable pageable);
+    List<Item> findByOwnerId(Long ownerId, Pageable pageable);
 
     @Query(" SELECT i FROM Item i " +
             "WHERE UPPER(i.name) LIKE UPPER(concat('%', ?1, '%')) " +
