+++ src/main/java/ru/practicum/shareit/item/service/ItemService.java
@@ -7,15 +7,15 @@ import ru.practicum.shareit.item.model.Comment;
 import java.util.List;
 
 public interface ItemService {
-    ItemDto create(ItemDto dto, Integer userId);
+    ItemDto create(ItemDto dto, Long userId);
 
-    ItemDto update(ItemDto dto, Integer itemId, Integer userId);
+    ItemDto update(ItemDto dto, Long itemId, Long userId);
 
-    ItemWithBookingInfoDto getById(Integer id);
+    ItemWithBookingInfoDto getById(Long id);
 
-    List<ItemDto> findAllByUser(Integer id);
+    List<ItemDto> findAllByUser(Long id);
 
     List<ItemDto> searchItems(String text, Integer from, Integer size);
 
-    Comment createComment(Comment comment, int itemId, int userId);
+    Comment createComment(Comment comment, Long itemId, Long userId);
 }
