+++ src/main/java/ru/practicum/shareit/request/repository/ItemRequestRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.shareit.request.ItemRequest;
 
 
-public interface ItemRequestRepository extends JpaRepository<ItemRequest, Integer> {
+public interface ItemRequestRepository extends JpaRepository<ItemRequest, Long> {
 }
